# backendApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Punch System API",
        "description": "打卡系統 API 文件",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/correction": {
            "get": {
                "tags": [
                    "Punch Correction"
                ],
                "summary": "取得個人補登紀錄",
                "description": "使用者可查詢自己的補登紀錄（可篩選日期）",
                "operationId": "getUserPunchCorrections",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "開始日期",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-03-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "結束日期",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-03-10"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功獲取補登紀錄",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PunchCorrection"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未授權"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Punch Correction"
                ],
                "summary": "提交補登請求",
                "description": "使用者提交補登請求",
                "operationId": "requestPunchCorrection",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "correction_type",
                                    "punch_time",
                                    "reason"
                                ],
                                "properties": {
                                    "correction_type": {
                                        "type": "string",
                                        "enum": [
                                            "punch_in",
                                            "punch_out"
                                        ],
                                        "example": "punch_in"
                                    },
                                    "punch_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-03-11 08:00:00"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "忘記打卡"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "補登申請成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PunchCorrection"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "打卡時間不能是未來時間"
                    },
                    "401": {
                        "description": "未授權"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/punch/correction/{id}/approve": {
            "put": {
                "tags": [
                    "Punch Correction"
                ],
                "summary": "審核補登申請（通過）",
                "description": "管理員審核通過補登申請",
                "operationId": "approvePunchCorrection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "補登請求 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "review_message": {
                                        "type": "string",
                                        "example": "審核通過"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "補登已通過審核",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PunchCorrection"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "此補登申請已被處理"
                    },
                    "401": {
                        "description": "未授權"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/punch/correction/{id}/reject": {
            "put": {
                "tags": [
                    "Punch Correction"
                ],
                "summary": "審核補登申請（拒絕）",
                "description": "管理員拒絕補登申請",
                "operationId": "rejectPunchCorrection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "補登請求 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "review_message"
                                ],
                                "properties": {
                                    "review_message": {
                                        "type": "string",
                                        "example": "補登原因不合理"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "補登請求已被拒絕",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PunchCorrection"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "請填寫拒絕原因"
                    },
                    "401": {
                        "description": "未授權"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PunchCorrection": {
                "title": "Punch Correction",
                "description": "打卡補登請求資料結構",
                "required": [
                    "id",
                    "user_id",
                    "correction_type",
                    "punch_time",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "correction_type": {
                        "type": "string",
                        "enum": [
                            "punch_in",
                            "punch_out"
                        ],
                        "example": "punch_in"
                    },
                    "punch_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-11 08:00:00"
                    },
                    "reason": {
                        "type": "string",
                        "example": "忘記打卡"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "approved",
                            "rejected"
                        ],
                        "example": "pending"
                    },
                    "review_message": {
                        "type": "string",
                        "example": "審核通過"
                    },
                    "approved_by": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "approved_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-11 09:00:00",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}